+++ src/modules/local_position_estimator/BlockLocalPositionEstimator.cpp
@@ -336,8 +336,8 @@ void BlockLocalPositionEstimator::update()
 	// if we have no lat, lon initialize projection to LPE_LAT, LPE_LON parameters
 	if (!_map_ref.init_done && (_estimatorInitialized & EST_XY) && _fake_origin.get()) {
 		map_projection_init(&_map_ref,
-				    _init_origin_lat.get(),
-				    _init_origin_lon.get());
+				    (double)_init_origin_lat.get(),
+				    (double)_init_origin_lon.get());
 
 		// set timestamp when origin was set to current time
 		_time_origin = _timeStamp;
